Address CodeRabbit review: fix animation loop on WebGL failure, stop captureWork on timeout, use image digest in CI, harden smoke assertions, extract navigateToRatePage helper, replace npx with pnpm exec
