+++ package.json
@@ -13,10 +13,10 @@
     "jiti": "^2.6.1"
   },
   "scripts": {
-    "lint:js": "sh -c 'if ls static/react/assets/*.js >/dev/null 2>&1 || ls static/assets/*.js >/dev/null 2>&1; then npx eslint static/react/assets/*.js static/assets/*.js; else echo \"No JavaScript files to lint\"; fi'",
-    "lint:html": "npx htmlhint app/templates/*.html",
+    "lint:js": "sh -c 'if ls static/react/assets/*.js >/dev/null 2>&1 || ls static/assets/*.js >/dev/null 2>&1; then pnpm exec eslint static/react/assets/*.js static/assets/*.js; else echo \"No JavaScript files to lint\"; fi'",
+    "lint:html": "pnpm exec htmlhint app/templates/*.html",
     "lint": "pnpm run lint:js && pnpm run lint:html",
-    "lint:fix": "sh -c 'if ls static/react/assets/*.js >/dev/null 2>&1 || ls static/assets/*.js >/dev/null 2>&1; then npx eslint static/react/assets/*.js static/assets/*.js --fix; else echo \"No JavaScript files to fix\"; fi'",
+    "lint:fix": "sh -c 'if ls static/react/assets/*.js >/dev/null 2>&1 || ls static/assets/*.js >/dev/null 2>&1; then pnpm exec eslint static/react/assets/*.js static/assets/*.js --fix; else echo \"No JavaScript files to fix\"; fi'",
     "dice:playground": "pnpm --filter frontend run dice:playground",
     "test:dice-playground": "pnpm --filter frontend run test:dice-playground",
     "test:dice-playground:update": "pnpm --filter frontend run test:dice-playground:update",
